chore: setup eslint and fix type check
