+++ app/admin/emotional-signature-engine/page.tsx
@@ -14,11 +14,12 @@ import { ScrollArea } from "@/components/ui/scroll-area"
 import { ChartContainer, ChartTooltip, ChartTooltipContent } from "@/components/ui/chart"
 import { BarChart, Bar, XAxis, YAxis, CartesianGrid, ResponsiveContainer, PieChart, Pie, Cell } from "recharts"
 import { createClient } from "@supabase/supabase-js"
+import { env } from "@/lib/env.client"
 import { Download, RefreshCw, Search, Send, X } from "lucide-react"
 
 // Initialize Supabase client
-const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL || ""
-const supabaseAnonKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY || ""
+const supabaseUrl = env.NEXT_PUBLIC_SUPABASE_URL || ""
+const supabaseAnonKey = env.NEXT_PUBLIC_SUPABASE_ANON_KEY || ""
 const supabase = createClient(supabaseUrl, supabaseAnonKey)
 
 // Emotion colors
