+++ components/vault/vault-solo-panel.tsx
@@ -21,11 +21,12 @@ import {
 } from "lucide-react"
 import { toast } from "sonner"
 import { createClient } from "@supabase/supabase-js"
+import { env } from "@/lib/env.client"
 
 // Supabase client setup
 const supabase = createClient(
-  process.env.NEXT_PUBLIC_SUPABASE_URL || "https://demo.supabase.co",
-  process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY || "demo-key",
+  env.NEXT_PUBLIC_SUPABASE_URL || "https://demo.supabase.co",
+  env.NEXT_PUBLIC_SUPABASE_ANON_KEY || "demo-key",
 )
 
 interface UserBalance {
@@ -86,7 +87,7 @@ export function VaultSoloPanel() {
 
       // Check if we have real Supabase credentials
       const hasRealCredentials =
-        process.env.NEXT_PUBLIC_SUPABASE_URL && process.env.NEXT_PUBLIC_SUPABASE_URL !== "https://demo.supabase.co"
+        env.NEXT_PUBLIC_SUPABASE_URL && env.NEXT_PUBLIC_SUPABASE_URL !== "https://demo.supabase.co"
 
       if (!hasRealCredentials) {
         // Use mock data when no real credentials
