+++ lib/ecosystem-health.ts
@@ -1,3 +1,5 @@
+import { env } from "@/lib/env.server"
+
 interface EcosystemHealthData {
   healthScore: number
   activeUsers: number
@@ -50,8 +52,8 @@ interface EcosystemHealthData {
 }
 
 export const fetchEcosystemHealth = async (): Promise<EcosystemHealthData> => {
-  const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL
-  const serviceRoleKey = process.env.SUPABASE_SERVICE_ROLE_KEY
+  const supabaseUrl = env.NEXT_PUBLIC_SUPABASE_URL
+  const serviceRoleKey = env.SUPABASE_SERVICE_ROLE_KEY
 
   if (!supabaseUrl || !serviceRoleKey) {
     throw new Error("Missing Supabase configuration")
