+++ lib/fetchLocaleHolidayData.ts
@@ -1,8 +1,9 @@
 "use client"
 
 import { createClient } from "@supabase/supabase-js"
+import { env } from "@/lib/env.client"
 
-const supabase = createClient(process.env.NEXT_PUBLIC_SUPABASE_URL!, process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!)
+const supabase = createClient(env.NEXT_PUBLIC_SUPABASE_URL!, env.NEXT_PUBLIC_SUPABASE_ANON_KEY!)
 
 export interface LocaleHoliday {
   id: string
