+++ lib/persona-cultural-adaptation.ts
@@ -1,8 +1,9 @@
 import { createClient } from "@supabase/supabase-js"
+import { env } from "@/lib/env.client"
 
 const supabase = createClient(
-  process.env.NEXT_PUBLIC_SUPABASE_URL || "https://demo.supabase.co",
-  process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY || "demo-key",
+  env.NEXT_PUBLIC_SUPABASE_URL || "https://demo.supabase.co",
+  env.NEXT_PUBLIC_SUPABASE_ANON_KEY || "demo-key",
 )
 
 export interface PersonaCulturalAdaptation {
