+++ middleware.ts
@@ -1,6 +1,7 @@
 import { createMiddlewareClient } from "@supabase/auth-helpers-nextjs"
 import { NextResponse } from "next/server"
 import type { NextRequest } from "next/server"
+import { env } from "@/lib/env.server"
 
 export async function middleware(req: NextRequest) {
   // Skip middleware for static files and API routes that don't need auth
@@ -8,8 +9,8 @@ export async function middleware(req: NextRequest) {
     req.nextUrl.pathname.startsWith("/_next") ||
     req.nextUrl.pathname.startsWith("/api/public") ||
     req.nextUrl.pathname.includes(".") ||
-    !process.env.NEXT_PUBLIC_SUPABASE_URL ||
-    !process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY
+    !env.NEXT_PUBLIC_SUPABASE_URL ||
+    !env.NEXT_PUBLIC_SUPABASE_ANON_KEY
   ) {
     return NextResponse.next()
   }
